Including web partt title in properties.searchableContent
