+++ Pzl.Part.CustomTextEditor/src/webparts/customTextEditor/ICustomTextEditorWebPartProps.ts
@@ -1,0 +1,10 @@
+import { TextBoxStyle } from "./components/TextBoxStyle";
+export interface ICustomTextEditorWebPartProps {
+    title: string;
+    Content: string;
+    searchableContent: string;
+    textBoxStyle: TextBoxStyle;
+    backgroundColor: string;
+    headerExpandColor: string;
+    underlineLinks: boolean;
+}
