+++ Pzl.Part.CustomTextEditor/src/webparts/customTextEditor/components/TextBoxStyle.tsx
@@ -1,0 +1,6 @@
+export enum TextBoxStyle {
+    WithBackgroundColor,
+    Accordion,
+    Regular,
+    RegularFade
+}
